+++ update-local.cjs
@@ -1,7 +1,6 @@
-const processes = require('node:child_process');
 const path = require('node:path');
 const fs = require('node:fs');
-const { common } = require('./update-common.cjs');
+const { common, clean } = require('./update-common.cjs');
 
 // if (process.versions.node.split('.')[0] != '16') {
 // 	console.error('Please use Node v16 for updating penguinmod versions!');
@@ -9,28 +8,10 @@ const { common } = require('./update-common.cjs');
 // }
 
 const pmDir = path.resolve(__dirname, 'penguinmod');
-try {
-  fs.accessSync(pmDir);
-  fs.rmSync(pmDir, {
-    recursive: true,
-    force: true,
-  });
-} catch {
-} finally {
-  fs.mkdirSync(pmDir);
-}
+clean(pmDir);
 
 const srcDir = path.resolve(__dirname, 'src');
-try {
-  fs.accessSync(srcDir);
-  fs.rmSync(srcDir, {
-    recursive: true,
-    force: true,
-  });
-} catch {
-} finally {
-  fs.mkdirSync(srcDir);
-}
+clean(srcDir);
 
 const pmSrc = path.resolve(__dirname, '../PenguinMod-Desktop-Gui');
 try {
@@ -47,11 +28,4 @@ fs.cpSync(pmSrc, pmDir, {
   force: true,
 });
 
-fs.rmSync(path.resolve(pmDir, 'package-lock.json'));
-processes
-  .spawn('npm i --force', {
-    cwd: pmDir,
-    stdio: 'inherit',
-    shell: true,
-  })
-  .once('close', common);
+common();